refactor: chat features
